+++ src/test/java/coupon/domain/DiscountMoneyPolicyTest.java
@@ -3,6 +3,7 @@ package coupon.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import coupon.domain.discountpolicy.DiscountMoneyPolicy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
